+++ joss/paper.bib
@@ -18,34 +18,46 @@
 }
 
 @article{kadapa2021simple,
-  title={A simple extrapolated predictor for overcoming the starting and tracking issues in the arc-length method for nonlinear structural mechanics},
-  author={Kadapa, Chennakesava},
-  journal={Engineering Structures},
-  volume={234},
-  pages={111755},
-  year={2021},
-  publisher={Elsevier}
+title = {A simple extrapolated predictor for overcoming the starting and tracking issues in the arc-length method for nonlinear structural mechanics},
+journal = {Engineering Structures},
+volume = {234},
+pages = {111755},
+year = {2021},
+issn = {0141-0296},
+doi = {https://doi.org/10.1016/j.engstruct.2020.111755},
+url = {https://www.sciencedirect.com/science/article/pii/S014102962034356X},
+author = {Chennakesava Kadapa},
+keywords = {Finite element analysis, Arc-length method, Structural stability, Limit points, Buckling}
 }
+
 @article{medina2020navigating,
-  title={Navigating the landscape of nonlinear mechanical metamaterials for advanced programmability},
-  author={Medina, Eder and Farrell, Patrick E and Bertoldi, Katia and Rycroft, Chris H},
-  journal={Physical Review B},
-  volume={101},
-  number={6},
-  pages={064101},
-  year={2020},
-  publisher={APS}
+  title = {Navigating the landscape of nonlinear mechanical metamaterials for advanced programmability},
+  author = {Medina, Eder and Farrell, Patrick E. and Bertoldi, Katia and Rycroft, Chris H.},
+  journal = {Phys. Rev. B},
+  volume = {101},
+  issue = {6},
+  pages = {064101},
+  numpages = {6},
+  year = {2020},
+  month = {Feb},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevB.101.064101},
+  url = {https://link.aps.org/doi/10.1103/PhysRevB.101.064101}
 }
 
 @article{lee2021elastic,
-  title={Elastic instabilities govern the morphogenesis of the optic cup},
-  author={Lee, Jeong-Ho and Park, Harold S and Holmes, Douglas P},
-  journal={Physical Review Letters},
-  volume={127},
-  number={13},
-  pages={138102},
-  year={2021},
-  publisher={APS}
+  title = {Elastic Instabilities Govern the Morphogenesis of the Optic Cup},
+  author = {Lee, Jeong-Ho and Park, Harold S. and Holmes, Douglas P.},
+  journal = {Phys. Rev. Lett.},
+  volume = {127},
+  issue = {13},
+  pages = {138102},
+  numpages = {7},
+  year = {2021},
+  month = {Sep},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevLett.127.138102},
+  url = {https://link.aps.org/doi/10.1103/PhysRevLett.127.138102}
 }
 
 @book{abaqus,
@@ -94,57 +106,70 @@ author = "D. Kamensky and Y. Bazilevs"
 
 @article{Mitusch2019, doi = {10.21105/joss.01292}, url = {https://doi.org/10.21105/joss.01292}, year = {2019}, publisher = {The Open Journal}, volume = {4}, number = {38}, pages = {1292}, author = {Sebastian K. Mitusch and Simon W. Funke and Jørgen S. Dokken}, title = {dolfin-adjoint 2018.1: automated adjoints for FEniCS and Firedrake}, journal = {Journal of Open Source Software} }
 
-@incollection{crisfield1981fast,
-  title={A fast incremental/iterative solution procedure that handles “snap-through”},
-  author={Crisfield, Michael A},
[…]
