+++ joss/paper.bib
@@ -24,7 +24,7 @@ volume = {234},
 pages = {111755},
 year = {2021},
 issn = {0141-0296},
-doi = {https://doi.org/10.1016/j.engstruct.2020.111755},
+doi = {10.1016/j.engstruct.2020.111755},
 url = {https://www.sciencedirect.com/science/article/pii/S014102962034356X},
 author = {Chennakesava Kadapa},
 keywords = {Finite element analysis, Arc-length method, Structural stability, Limit points, Buckling}
@@ -100,7 +100,7 @@ volume = "344",
 pages = "477--498",
 year = "2019",
 issn = "0045-7825",
-doi = "https://doi.org/10.1016/j.cma.2018.10.002",
+doi = "10.1016/j.cma.2018.10.002",
 author = "D. Kamensky and Y. Bazilevs"
 }
 
@@ -114,7 +114,7 @@ number = {1},
 pages = {55-62},
 year = {1981},
 issn = {0045-7949},
-doi = {https://doi.org/10.1016/0045-7949(81)90108-5},
+doi = {10.1016/0045-7949(81)90108-5},
 url = {https://www.sciencedirect.com/science/article/pii/0045794981901085},
 author = {M.A. Crisfield}
 }
@@ -127,7 +127,7 @@ volume = {77},
 number = {9},
 pages = {1290-1321},
 keywords = {arc-length control, path-following technique, solution control, energy release rate},
-doi = {https://doi.org/10.1002/nme.2447},
+doi = {10.1002/nme.2447},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/nme.2447},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/nme.2447},
 year = {2009}
@@ -141,7 +141,7 @@ number = {7},
 pages = {529-551},
 year = {1979},
 issn = {0020-7683},
-doi = {https://doi.org/10.1016/0020-7683(79)90081-7},
+doi = {10.1016/0020-7683(79)90081-7},
 url = {https://www.sciencedirect.com/science/article/pii/0020768379900817},
 author = {E. Riks}
 }
@@ -167,7 +167,7 @@ number = {1},
 pages = {79-116},
 year = {1986},
 issn = {0045-7825},
-doi = {https://doi.org/10.1016/0045-7825(86)90079-4},
+doi = {10.1016/0045-7825(86)90079-4},
 url = {https://www.sciencedirect.com/science/article/pii/0045782586900794},
 author = {J.C. Simo and L. Vu-Quoc}
 }
@@ -189,7 +189,7 @@ volume = {9},
 pages = {107-111},
 year = {2019},
 issn = {2352-7110},
-doi = {https://doi.org/10.1016/j.softx.2018.10.005},
+doi = {10.1016/j.softx.2018.10.005},
 url = {https://www.sciencedirect.com/science/article/pii/S2352711018300979},
 author = {Miguel A. Rodriguez and Christoph M. Augustin and Shawn C. Shadden},
 keywords = {Computational mechanics, Fluid mechanics, Solid mechanics, Finite element method}
@@ -202,7 +202,7 @@ journal = {International Journal for Numerical Methods in Engineering},
 volume = {14},
 number = {8},
 pages = {1262-1267},
-doi = {https://doi.org/10.1002/nme.1620140811},
+doi = {10.1002/nme.1620140811},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/nme.1620140811},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/nme.1620140811},
 year = {1979}
[…]
