### 2023-08-01 #5604 @w-luo-academic
Could you please add my new email address: ***@***.***?
My Northwestern email was no longer being used for quite a while.

Thanks,
Wen

On Tue, Aug 1, 2023 at 11:37 AM The Open Journals editorial robot <
***@***.***> wrote:

> Assigned #5604 <https://github.com/openjournals/joss-reviews/issues/5604>
> to @w-luo-academic <https://github.com/w-luo-academic>.
>
