### 2023-08-01 #5604 @kingyin3613
> Hi @kingyin3613, would you be interested in reviewing this JOSS submission? You can read the [draft](https://github.com/openjournals/joss-papers/blob/joss.05604/joss.05604/10.21105.joss.05604.pdf) and also check out their [software repository](https://github.com/pprachas/fenics_arclength).
>
> The JOSS review process is detailed in this [link](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). I will be happy to help with other questions. Let me know if this interests you.

Yes, I am willing to review this interesting package!
