### 2023-08-01 #5727 @editorialbot
[REVIEW]: FEniCS-arclength: A numerical continuation package in FEniCS for nonlinear problems in solid mechanics
**Submitting author:** @pprachas (<a href="http://orcid.org/0009-0000-3325-1410">Peerasait Prachaseree</a>)
**Repository:** https://github.com/pprachas/fenics_arclength
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
