### 2023-08-10 #5727 @kingyin3613
Hi @pprachas, very useful and well-documented package! While I am trying to install this package on various OS (Windows + Mac + Ubuntu), I noticed two minor issues:
1. have you ever considered using automated package management for easy installation? For example, since your package and FEniCS are mainly written in Python, PyPI (classical "pip install") can be used, or in conda environment, Conda-forge can be used. Using automated package management would simplify the installation for users under very different environments, and would benefit package maintenance in cases where the dependencies are changed.
2. in your documentation, I could not find a clear guide for contributing to the package/reporting bugs/etc. I suggest you add a community support or a contributing section. You may take a look at these packages as references: https://github.com/bessagroup/CRATE or http://gala.adrian.pw/en/latest/contributing.html.

I am looking for your reply!
