### 2023-08-15 4b6e02b
added pyproject.toml; edited notebooks and scripts
