+++ fenics_arclength.egg-info/SOURCES.txt
@@ -1,0 +1,11 @@
+LICENSE
+README.md
+pyproject.toml
+arc_length/__init__.py
+arc_length/displacement_control_solver.py
+arc_length/force_control_solver.py
+arc_length/rotation_parametrization.py
+fenics_arclength.egg-info/PKG-INFO
+fenics_arclength.egg-info/SOURCES.txt
+fenics_arclength.egg-info/dependency_links.txt
+fenics_arclength.egg-info/top_level.txt

[…]
