+++ README.md
@@ -23,10 +23,11 @@ More information on the arc-length method and the solution approach can be found
 
 ## Table of Contents
 * [Dependencies](#dependencies)
-* [Usage](#usage)
+* [Installation](#installation)
 * [Contents in this Repository](#contents)
 * [Validation](#validation)
 * [Theory](#theory)
+* [Community Contributions](#community)
 
 
 ## Dependencies <a name="dependencies"></a>
@@ -86,24 +87,19 @@ Note that in this case the conda environment name will be ``fenics_arclength``.
 
 **Note: For Docker installation, numpy, scipy, and matplotlib should be installed alongside FEniCS by default. To enable jupyter notebooks for FEniCS, please see this link: [Jupyter notebooks for Docker installation of FEniCS](https://fenics.readthedocs.io/projects/containers/en/latest/jupyter.html). However, the scripts in the ``validation`` can be run without Jupyter notebooks.**
  
-## Usage
-To use this arc-length solver, download and append this repository to the python path. Two common methods to do this are:
+## Installation
+Once FEniCS has been installed, our package can easily be install through pip with the following command:
 
-1. Add directory to `PYTHONPATH`
+    pip install git+https://github.com/pprachas/fenics_arclength.git@master
 
-- On Ubuntu or Mac:
+In cases where the whole github repository is needed, the github repo can first be cloned before installation:
 
-        export PYTHONPATH=<path/to/fenics_arclength>:$PYTHONPATH
-  Note that you might have to first find your python path in the computer and add it to the bash profile first. 
- 
-- On Windows:
+    git clone https://github.com/pprachas/fenics_arclength.git
+    cd fenics_arclength
+    pip install .
 
-        set PYTHONPATH=<path/to/fenics_arclength>;%PYTHONPATH%
+In the case of developer's version, the last line can be replaced with ``pip install -e .``
 
-2. Append directory to path in python script:
-
-        import sys
-        sys.path.append('path/to/fenics_arclength')
 
 
 ## Contents in this repository <a name="contents"></a>
@@ -113,10 +109,10 @@ To use this arc-length solver, download and append this repository to the python
 |[arc_length](arc_length)| contains the code for our arc-length implementation; both force and displacement control solvers are contained there |
 |[docs](docs)| the build files for our [readthedocs documentation](https://fenics-arclength.readthedocs.io/en/latest/) |
 |[examples](examples) | contains Jupyter notebook examples to use our arc-length implementation. Note that Jupyter notebooks has to be installed in the FEniCS environment for the notebooks to run. | 
-|[validation](validation)| contains python scripts to compare our solver with analytical solutions/solutions in literature. <br><br> To run the scripts run: <br> <br> ``python3 validation/validate_xx.py`` <br><br> from the project root directory. More information in section [Validation](#validation) |
+|[validation](validation)| contains python scripts to compare our solver with analytical solutions/solutions in literature. |
 
 ## Validation <a name="validation"></a>
-To validate that our arc-length solver works we provide 3 validation examples. To run the examples go to the root directory and run ``python3 validation/validate_xx.py``. The output plots will be in the ``validation/plots`` folder. The visualize the Paraview files, append ``-p`` at the end of the command line (i.e. ``python3 validation//validate_xx.py -p``). The available python scripts are:
+To validate that our arc-length solver works we provide 3 validation examples. To run the examples go to the ``validation`` directory and run ``python3 validate_xx.py``. The output plots will be in the ``plots`` folder. The visualize the Paraview files, append ``-p`` at the end of the command line (i.e. ``python3 validate_xx.py -p``). The available python scripts are:
 * ``validate_3Dbeamsmall.py``
[…]
