### 2023-08-16 6b39549
Added pip installation instructions to docs
