+++ docs/source/index.rst
@@ -37,6 +37,49 @@ For conda installations (available only on Mac and Linux), we also provide an en
 
 Package Installation
 ####################
+To use the arc-length solver, you can install it using pip or by downloading and adding the repository to your Python path.
+
+Install using pip
+*****************
+
+The recommended way to install the arc-length solver is to use pip. To do this, run the following command in your terminal:
+
+.. code-block:: bash
+
+   pip install git+https://github.com/pprachas/fenics_arclength.git@master
+
+This will install the latest version of the arc-length solver from the master branch of the repository.
+
+Install by downloading and adding to Python path
+************************************************
+
+If you prefer to download and install the arc-length solver manually, you can do so by following these steps:
+
+1. Clone the repository to your local machine:
+   
+   .. code-block:: bash
+
+      git clone https://github.com/pprachas/fenics_arclength.git
+
+2. Change directory to the fenics_arclength directory:
+   
+   .. code-block:: bash
+
+      cd fenics_arclength
+
+3. Install the arc-length solver:
+
+   .. code-block:: bash
+
+      pip install .
+
+This will install the arc-length solver from the source code in the fenics_arclength directory.
+
+Once the arc-length solver is installed, you can import it into your Python scripts and use it to solve arc-length problems.
+
+Install by downloading and adding to Python path (Legacy)
+*********************************************************
+
 To use our arc-length solver, download and append our `repository <https://github.com/pprachas/fenics_arclength>`_ to the python path. Common methods to do this are:
 
 * Add directory to ``PYTHONPATH``:
