### 2023-08-16 #5727 @pprachas
Hi @kingyin3613,

Thanks for your response! In an attempt to address the issues you presented, we made the following changes:

1. We made our package pip installable. Instructions on the specific commands are now on readthedocs and GitHub README. However, a working FEniCS installation is required. There are many ways to install FEniCS and we specified the common ones in our GitHub README and provide links to the official FEniCS installation documentation. We note that this is inline with most FEniCS-based packages in JOSS (i.e. [dolfin-adjoint](https://github.com/dolfin-adjoint/pyadjoint), [pulse](https://github.com/finsberg/pulse)).

2. We added a guide on community contribution in the last section of the README.

Please let us know what you think!
