### 2023-08-27 #5727 @kingyin3613
Hi @pprachas,

Your new pip installation looks good! However, I found this issue: https://github.com/pprachas/fenics_arclength/issues/10 when installing fenics_arclength in Windows+Docker environment. I am looking for your solutions!

The contribution section should be good to go according to JOSS requirements!
