### 2023-09-07 85d39ee
Update math_prelim.rst
