### 2023-09-08 54e1643
Update math_prelim.rst
