+++ README.md
@@ -244,7 +244,7 @@ C^\top K_T C & C^\top K \mathbf{u}_p \\
 \end{bmatrix}
 ```
 
-Similar to Ref. 3 and Ref. 4, we used the Shur complement to solve the system of equations. For more details refer to the Ref 3 and Ref 4. 
+Similar to Ref. 3 and Ref. 4, we solve the block system of equations by part using the Sherman–Morrison formula. For more details refer to the Ref 3 and Ref 4. 
 
 ## Community Contributions
 
