### 2023-09-15 ada060f
Update math_prelim.rst
