+++ pyproject.toml
@@ -1,7 +1,5 @@
 [build-system]
 requires = ["setuptools>=61.0.0", "wheel"]
-build-backend = "setuptools.build_meta"
-
 
 [project]
 name = "fenics-arclength"
