+++ pyproject.toml
@@ -12,7 +12,7 @@ authors = [
 description = "An arc-length Riks solver implemented with FEniCS"
 readme = "README.md"
 license = { file="LICENSE" }
-requires-python = ">=3.6"
+requires-python = ">=3.8"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
