+++ .readthedocs.yaml
@@ -1,6 +1,11 @@
 version: 2
 formats: all
 
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.11"
+    
 formats:
   - htmlzip
   
