+++ .readthedocs.yaml
@@ -5,11 +5,11 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.11"
-    
+
+python:
+  version: 3.11
+  install:
+    - requirements: docs/requirements.txt
 formats:
   - htmlzip
   
-python:
-  version: 3
-  install:
-    - requirements: docs/requirements.txt
