+++ .readthedocs.yaml
@@ -7,7 +7,7 @@ build:
     python: "3.11"
 
 python:
-  version: 3.11
+  version: "3.11"
   install:
     - requirements: docs/requirements.txt
 formats:
