+++ pyproject.toml
@@ -1,5 +1,6 @@
 [build-system]
-requires = ["setuptools>=61.0.0", "wheel"]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
 
 [project]
 name = "fenics-arclength"
