+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools"]
+requires = ["setuptools", "wheels"]
 build-backend = "setuptools.build_meta"
 
 [project]
