+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools", "wheels"]
+requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [project]
