### 2023-09-19 pprachas/fenics_arclength#10 @pprachas
Hi @kingyin3613,
Thank you for your patience while we worked to fix the issue.
 I think we now fixed the issue with pip installation!

  Can you please retry this command in the Docker installation?

  ``pip install git+https://github.com/pprachas/fenics_arclength.git@master``

  Please let us know how it goes!
