### 2023-09-23 pprachas/fenics_arclength#10 @kingyin3613
@pprachas now it works! However, I have to use the option `--user` to install, like

> pip install --user git+https://github.com/pprachas/fenics_arclength.git@master

otherwise, it yields another issue:
```
Collecting git+https://github.com/pprachas/fenics_arclength.git@master
  Cloning https://github.com/pprachas/fenics_arclength.git (to revision master) to /tmp/pip-req-build-44e4jm6d
  Running command git clone -q https://github.com/pprachas/fenics_arclength.git /tmp/pip-req-build-44e4jm6d

Building wheels for collected packages: fenics-arclength
  Building wheel for fenics-arclength (setup.py) ... done
  Stored in directory: /tmp/pip-ephem-wheel-cache-vnqk2vaa/wheels/59/70/b2/d01978ff11be4ca177d9dfce513a04a3aa1ca765b59f9171de
Successfully built fenics-arclength
Installing collected packages: fenics-arclength
ERROR: Could not install packages due to an EnvironmentError: [Errno 13] Permission denied: '/usr/local/lib/python3.6/dist-packages/fenics_arclength-0.1.0.dist-info'
Consider using the `--user` option or check the permissions.
```

FYI.
