### 2023-10-12 pprachas/fenics_arclength#12 @kingyin3613
No module named dolfin
Hi, I'm still trying to install fenics_arclength in Windows+Docker environment through the following commands:

> docker pull quay.io/fenicsproject/stable:current
> docker run -ti quay.io/fenicsproject/stable:current
> pip install --user git+https://github.com/pprachas/fenics_arclength.git@master

It turned out that I still could not run any fenics_arclength examples and validations using pip install, hence I switched to clone and install like:

> docker pull quay.io/fenicsproject/stable:current
> docker run -ti quay.io/fenicsproject/stable:current
> git clone https://github.com/pprachas/fenics_arclength.git
> cd fenics_arclength
> pip install --user .
> cd validation
> python validate_3Dbeamsmall.py

An error occurred:
```
Traceback (most recent call last):
  File "validate_bilayer.py", line 1, in <module>
    from dolfin import *
ImportError: No module named dolfin
```

I tried to fix the problem with these solutions: https://fenicsproject.discourse.group/t/modulenotfounderror-no-module-named-dolfin/11427/2  and  https://fenicsproject.discourse.group/t/no-module-named-dolfin-ubuntu-20-04/4528.
However, none of these worked.

Then I tried to use Mac OS to install fenics-arclength by:

> git clone https://github.com/pprachas/fenics_arclength.git
> cd fenics_arclength
> pip install --user .
> cd validation
> python validate_3Dbeamsmall.py

it raised the same error:
```
Traceback (most recent call last):
  File "validate_bilayer.py", line 1, in <module>
    from dolfin import *
ImportError: No module named dolfin
```

I'm thinking of an alternative solution, could you please provide a more detailed step-by-step installation to example-run guidance for users who never used Fenics before? That would be very helpful!

Hao
