### 2023-10-12 #5727 @kingyin3613
Hi @pprachas,

I'm still stuck at installation and opened a new issue: https://github.com/pprachas/fenics_arclength/issues/12, this issue happened for both Windows+Docker and Mac OS environments.

Hao
