### 2023-10-12 pprachas/fenics_arclength#12 @pprachas
Hi @kingyin3613

As a quick debug you can check the python version by ``python --version``
If your default python version is is python 2 then FEniCS is not supported.

In this case you will need to use ``python3`` command to run scripts (i.e. ``python3 validate_3Dbeamsmall.py``) -- we have already specified this in the instruction of running the validation scripts. We also recommend running the beam examples first as they are faster than the bilayer examples.

Please let us know if this works! If not, we will take another look!
