### 2023-10-12 pprachas/fenics_arclength#12 @kingyin3613
Hi @pprachas, `python3 validate_3Dbeamsmall.py` worked! Thanks for your prompt rely. I will proceed to check the results soon!
