### 2023-10-13 #5727 @kingyin3613
Hi @prashjha and @pprachas, this software significantly supplements the open-source finite element FEniCS in the nonlinear equations regime. The documentation is well-written and complete, I could successfully install the package by following the README step-by-step, and all validation tests for classical nonlinear FEM benchmarks have been passed, results are correct. All boxes are ticked, I recommend acceptance in JOSS!
