### 2023-10-29 #5727 @w-luo-academic
Hi @pprachas,

While running the validation examples, I encountered the following message:
"line 4, in <module> from ufl import diag, Jacobian, shape
ModuleNotFoundError: No module named 'ufl'

Wen

On Fri, Oct 13, 2023 at 11:17 AM Prashant K. Jha ***@***.***>
wrote:

> Hi @kingyin3613 <https://github.com/kingyin3613>, thank you for your
> efforts and for providing your recommendations. I appreciate it.
>
