### 2023-10-29 #5727 @pprachas
Hi @w-luo-academic ,

Thanks for bringing this issue up! To hekp us debug the issue, how did you install FEniCS(i.e. conda/docker) and what is your computer's operating system?
