### 2023-10-29 #5727 @w-luo-academic
I use the wsl in windows.

On Sun, Oct 29, 2023 at 3:53 PM Peerasait Prachaseree (Jeffrey) <
***@***.***> wrote:

> Hi @w-luo-academic <https://github.com/w-luo-academic> ,
>
> Thanks for bringing this issue up! To hekp us debug the issue, how did you
> install FEniCS(i.e. conda/docker) and what is your computer's operating
> system?
>
