### 2023-10-29 #5727 @w-luo-academic
I simply followed the following steps posted on the repo:

    sudo apt-get install software-properties-common
    sudo add-apt-repository ppa:fenics-packages/fenics
    sudo apt-get update
    sudo apt-get install fenics

On Sun, Oct 29, 2023 at 4:08 PM Peerasait Prachaseree (Jeffrey) <
***@***.***> wrote:

> Couple more questions -- please bear with me.
>
>    1. Did you install FEniCS directly with PPA or with a conda
>    environment?
>    2. If you installed FEniCS with PPA is there other software in wsl
>    that also calls ufl? It might be due to a conflict of ufl versions (see post
>    by fenics developers here
>    <https://fenicsproject.discourse.group/t/announcement-ufl-legacy-and-legacy-dolfin/11583>
>    ).
>
> Please let us know if this is the issue!
>
