### 2023-10-29 #5727 @pprachas
Can you please run these commands in wsl and show us the output?

``python3 -c "import ufl; import dolfin; print(f'dolfin version: {dolfin.__version__}'); print(f'ufl version: {ufl.__version__}')"``
