### 2023-10-29 #5727 @w-luo-academic
Following is the result:

Traceback (most recent call last):

  File "<string>", line 1, in <module>

ModuleNotFoundError: No module named 'ufl'

On Sun, Oct 29, 2023 at 4:42 PM Peerasait Prachaseree (Jeffrey) <
***@***.***> wrote:

> Can you please run these commands in wsl and show us the output?
>
> python3 -c "import ufl; import dolfin; print(f'dolfin version:
> {dolfin.__version__}'); print(f'ufl version: {ufl.__version__}')"
>
