### 2023-10-29 #5727 @pprachas
Thanks! Again please bear with me as I try to pinpoint the issue. Can you please run this command (theres a slight change in the import)?

`` python3 -c "import ufl_legacy as ufl; import dolfin; print(f'dolfin version: {dolfin.__version__}'); print(f'ufl version: {ufl.__version__}')"``
