### 2023-12-11 #5727 @w-luo-academic
Hi Peerasait,

Sorry for the late reply, it was a crazy semester. I will do it by this
week!

Wen

On Mon, Dec 11, 2023 at 11:27 AM Peerasait Prachaseree (Jeffrey) <
***@***.***> wrote:

> Hi @w-luo-academic <https://github.com/w-luo-academic>,
>
> Are there any updates on running the software or anything we can help with?
>
