### 2024-01-14 #5727 @w-luo-academic
Hi Jeffrey,

I was trying to install FENICS on my Windows PC - it was a pain in the
neck. Is there any good tutorial on installation?

Thanks
Wen

On Fri, Jan 12, 2024 at 2:43 PM Peerasait Prachaseree (Jeffrey) <
***@***.***> wrote:

> Hi @w-luo-academic <https://github.com/w-luo-academic>,
>
> First of all Happy New year! Hope you are having a good start to the year.
>
> I am wondering if there are any updates in the review or anything we can
> help with?
>
> Thanks!
>
