### 2024-01-14 #5727 @pprachas
Hi @w-luo-academic,

If you are using wsl the simplest method is to create a conda environment and install fenics and fenics-arclength from conda-forge.

Here are link to install [miniconda](https://docs.conda.io/projects/miniconda/en/latest/) (run the commands Linux in wsl terminal; you might have to restart wsl after installation).

After install miniconda you can install fenics (and all the dependencies for fenics-arclength) with the provided [environment.yml](https://github.com/pprachas/fenics_arclength/blob/master/environment.yml) in our github repo. To use the ``environment.yml`` navigate to the directory of the downloaded file and use the following commands:

```
conda create -n fenicsproject
conda activate fenicsproject
conda env update -f environment.yml
```
Jupyter notebooks are optional but are needed to run examples (Jupyter notebooks can be install by ``conda install -c conda-forge jupyter``)

Finally, to install fenics-arclength and run all the examples we recommend you first clone the repository then install fenics-arclength:

```
git clone https://github.com/pprachas/fenics_arclength.git
cd fenics_arclength
pip install .
```

Hopefully this clears up installation issues and let me know if you have anymore questions! We will review the installation instructions and make sure that it is clear in the meantime.
