### 2024-01-20 2e26024
Update examples and validation to deal with ufl_legacy
