+++ validation/validate_3Dbeamlarge.py
@@ -1,12 +1,17 @@
 from dolfin import *
 import numpy as np
 import matplotlib.pyplot as plt
-from ufl import diag, Jacobian, shape
 import sys
 from arc_length.rotation_parametrization import ExponentialMap # import rotation parameterization for 3D beams
 from arc_length.force_control_solver import force_control # import force control formulation of arc-length solver
 from pathlib import Path
 import argparse
+
+# Dealing with ufl legacy
+try:
+    from ufl import diag, Jacobian, shape
+except:
+    from ufl_legacy import diag, Jacobian, shape
 
 parameters["form_compiler"]["cpp_optimize"] = True
 parameters["form_compiler"]["quadrature_degree"] = 3
