+++ docs/requirements.txt
@@ -1,3 +1,4 @@
-myst-nb
-sphinx_design
-sphinx_book_theme
+Sphinx=4.5.0
+myst-nb=0.17.1
+sphinx_design=0.3.0
+sphinx_book_theme=0.3.3
