### 2024-01-21 292c269
Updated Sphinx requirements
