+++ .readthedocs.yaml
@@ -4,7 +4,7 @@ formats: all
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.11"
+    python: "3.8.10"
 
 python:
   install:
