### 2024-01-22 51a273a
Added google colab functionalities to examples and validation
