### 2024-01-24 #5727 @prashjha
@pprachas, reviews are almost complete. I will read the draft once and let you know if I have any comments.

Meanwhile, could you please archive (if not done already) the release using zenodo and provide the archive reference so that I can associate it with your JOSS submission? Make sure that the zenodo archive's title matches this JOSS article's title.

If you have updated a version of your code, let me know, and I can update it here.
