### 2024-01-24 #5727 @w-luo-academic
I can no longer make any changes to the checklist. Is there a way to
reactivate it?

On Wed, Jan 24, 2024 at 8:37 AM Prashant K. Jha ***@***.***>
wrote:

> @w-luo-academic <https://github.com/w-luo-academic>, one last thing,
> could you please complete your checklist
> <https://github.com/openjournals/joss-reviews/issues/5727#issuecomment-1712675366>?
> If you feel some items could not be marked complete, could you please
> provide a reason (for only major ones)? Thank you!!
>
