### 2024-01-24 #5727 @w-luo-academic
That is fine. Please check the remaining boxes. Once FEniCS is installed,
the validation is very straightforward. It was a nightmare to install it on
Windows though.

On Wed, Jan 24, 2024 at 10:51 AM Prashant K. Jha ***@***.***>
wrote:

> weird..I just checked the remaining items in "General checks". Let me know
> if it is okay. If you like I can check other remaining items for you.
> Please confirm.
>
