### 2024-01-26 f856696
Fixed wording, added section about Colab
