+++ joss/paper.md
@@ -39,7 +39,7 @@ Such structural instabilities are induced by geometric nonlinearities, whereas m
 
 FEniCS [@alnaes2015fenics; @logg2012automated] is an open-source software that solves partial differential equations (PDEs) through finite element analysis (FEA). Building on the versatile FEniCS platform, researchers have created both detailed tutorials [@bleyer2018numericaltours] and software packages [@finsberg2019pulse; @rodriguez2019fenics] to facilitate research in specific application domains. For example, there are multiple open-source software libraries designed to increase the existing capabilities of FEniCS [@farrell2016computation; @Kamensky2019; @Mitusch2019]. Building on this established precedent, we have implemented our arclength solver as an add-on library on top of FEniCS, and also demonstrated usage of our solver via multiple documented examples involving nonlinear continuum and beam finite elements at large deformations. Notably, our intent is to keep the solver usage similar to already established FEniCS solvers to enable off-the-shelf implementation within FEniCS workflows. To this end, our arclength solver can be integrated into existing FEniCS code that is based on a Newton solver just by changing the choice of solver. Furthermore, our code and documentation are organized such that more advanced users can make custom modifications as needed, for example pinpointing and branch switching algorithms [@wriggersQuadraticallyConvergentProcedure1988]. 
 
-In brief, our implementation follows a recently published method [@kadapa2021simple] adopting an extrapolated predictor scheme using the previously converged two solution steps, and solves the augmented arclength matrix equation as the corrector scheme. The main advantage of the extrapolation prediction scheme is in its ability to determine the forward direction of the equilibrium path without doing extra computations such as evaluating the determinant of the stiffness matrix [@crisfield1981fast], although calculating the sign of the determinant can be simple depending on the type of linear solver (i.e., the sign of the determinant can be extracted during numerical factorization). For displacement-controlled arclength problems, similar to a multifreedom constraint problem, we introduced a constraint matrix to augment the prescribed non-zero Dirichlet boundary conditions and the free degrees of freedoms to solve the arclength equations [@batoz1979incremental; @verhoosel2009dissipation]. More details about the theory, implementation, and usage of our arclength solver and general mathematical preliminaries can be found in our extensive ReadTheDocs documentation (<https://fenics-arclength.readthedocs.io/en/latest/>). 
+In brief, our implementation follows a recently published method [@kadapa2021simple] adopting an extrapolated predictor scheme using the previously converged two solution steps, and solves the augmented arclength matrix equation as the corrector scheme. The main advantage of the extrapolation prediction scheme is in its ability to determine the forward direction of the equilibrium path without doing extra computations such as evaluating the determinant of the stiffness matrix [@crisfield1981fast], although calculating the sign of the determinant can be simple depending on the type of linear solver (i.e., the sign of the determinant can be extracted during numerical factorization). For displacement-controlled arclength problems, similar to a multifreedom constraint problem, we introduced a constraint matrix to augment the prescribed non-zero Dirichlet boundary conditions and the free degrees of freedoms to solve the arclength equations [@batoz1979incremental; @verhoosel2009dissipation]. More details about the theory, implementation, usage of our arclength solver, and general mathematical preliminaries can be found in our extensive ReadTheDocs documentation (<https://fenics-arclength.readthedocs.io/en/latest/>). 
 
 # Statement of Need
[…]
