+++ joss/paper.bib
@@ -282,3 +282,13 @@ pages = {15-41},
 year = {2001},
 doi = {10.1137/S0895479899358194},
 }
+
+@misc{Google-Colab,
+  title        = {Google Colaboratory},
+  howpublished = {\url{https://colab.google/}},
+}
+
+@misc{FEM-on-COlab,
+  title        = {FEM on Colab},
+  howpublished = {\url{https://fem-on-colab.github.io/index.html}},
+}
