+++ joss/paper.bib
@@ -285,10 +285,10 @@ doi = {10.1137/S0895479899358194},
 
 @misc{Google-Colab,
   title        = {Google Colaboratory},
-  howpublished = {\url{https://colab.google/}},
+  url = https://colab.google/
 }
 
-@misc{FEM-on-COlab,
+@misc{FEM-on-Colab,
   title        = {FEM on Colab},
-  howpublished = {\url{https://fem-on-colab.github.io/index.html}},
+  url = https://fem-on-colab.github.io/index.html
 }
