### 2024-01-26 64ad04a
Corrected bibtex file
