+++ joss/paper.bib
@@ -285,7 +285,7 @@ doi = {10.1137/S0895479899358194},
 
 @misc{Google-Colab,
   title        = {Google Colaboratory},
-  url = https://colab.google/
+  url = https://colab.google
 }
 
 @misc{FEM-on-Colab,
