### 2024-01-26 697c9ad
Modify bibtex file
