+++ joss/paper.bib
@@ -285,10 +285,10 @@ doi = {10.1137/S0895479899358194},
 
 @misc{Google-Colab,
   title        = {Google Colaboratory},
-  url = https://colab.google
+  url = {https://colab.google/}
 }
 
 @misc{FEM-on-Colab,
   title        = {FEM on Colab},
-  url = https://fem-on-colab.github.io/index.html
+  url = {https://fem-on-colab.github.io/index.html}
 }
