+++ joss/paper.bib
@@ -286,11 +286,11 @@ doi = {10.1137/S0895479899358194},
 @misc{Google-Colab,
   title        = {Google Colaboratory},
   url = {https://colab.google/}, 
-  year = 2024
+  year = {2024}
 }
 
 @misc{FEM-on-Colab,
   title        = {FEM on Colab},
   url = {https://fem-on-colab.github.io/index.html}
-  year = 2024
+  year = {2024}
 }
