+++ joss/paper.bib
@@ -291,6 +291,6 @@ doi = {10.1137/S0895479899358194},
 
 @misc{FEM-on-Colab,
   title        = {FEM on Colab},
-  url = {https://fem-on-colab.github.io/index.html}
+  url = {https://fem-on-colab.github.io/index.html},
   year = {2024}
 }
