### 2024-01-29 #5727 @prashjha
@pprachas, could you take a look at your joss paper markdown file and address the above errors? I think the first error is related to '\S' in the table. The second error is possibly due to redefining the tag. I can take a closer look if you want. Let me know once you are done.
