### 2024-01-29 #5727 @pprachas
Thank you @kingyin3613 and @w-luo-academic for your reviews! I understand FEniCS can be frustrating to work with sometimes so I really appreciate both of you going through the code and all the useful comments!

@prashjha I think I addressed all the errors above -- Let me know if the issue is fixed or of there is anything else for me to do.
