### 2024-02-01 #5727 @prashjha
@pprachas, there is still an error about labeling. Can you try removing `{#tab:examples}` from [line 53](https://github.com/pprachas/fenics_arclength/blob/78a7cb6001ec146ce2db6d39571ec3a137fa2c98/joss/paper.md?plain=1#L53)? If I am not wrong, tab:examples label is defined in [line 53](https://github.com/pprachas/fenics_arclength/blob/78a7cb6001ec146ce2db6d39571ec3a137fa2c98/joss/paper.md?plain=1#L53) and then in [line 86](https://github.com/pprachas/fenics_arclength/blob/78a7cb6001ec146ce2db6d39571ec3a137fa2c98/joss/paper.md?plain=1#L86).
