+++ joss/paper.md
@@ -50,7 +50,6 @@ Materials and structures that undergo large deformation often exhibit fascinatin
 
 In \autoref{fig:leesframe}, we show an example of our arclength implementation for solving the Lee's frame problem [@lee1968large], a popular benchmark problem in nonlinear structural mechanics. In this problem, the frame is discretized using 2D Simo-Reissner beams [@simo1986three] with mixed elements. The elements have linear Lagrange interpolation functions for both displacement and rotation, and use reduced integration to avoid locking. \autoref{fig:leesframe}a gives an overview of the Lee's frame problem and illustrates the equilibrium path (\autoref{fig:leesframe}b) and the deformed configuration (\autoref{fig:leesframe}c) in comparison with the literature [@kadapa2021simple]. The Lee's frame example demonstrates that our solver is able to trace unstable states, where traditional Newton solvers would fail. Additional validation examples on problems in nonlinear structural mechanics with beam elements (compared to solutions from Kadapa et al. [@kadapa2021simple]), and bilayer wrinkling with continuum elements (compared to analytical solutions [@allensandwich; @budday2017wrinkling; @cao2012wrinkling]) can be found in the examples folder of the Github Repository (<https://github.com/pprachas/fenics_arclength/tree/master/examples>), and in the ReadTheDocs (<https://fenics-arclength.readthedocs.io/en/latest/index.html#notebook-examples>) documentation. These examples can be easily run through Google Colab [@Google-Colab] with FEM on Colab [@FEM-on-Colab]. Critically, we demonstrated that our solver is able to trace the complex equilibrium paths of popular benchmark problems such as $215^\circ$ and $180^\circ$ archs (see \autoref{tab:examples} for all the provided examples in our documentation). In addition to validation problems, our documentation also contains the solver's API, example usage of our solver for continuum problems, and example usage of our displacement controlled arclength scheme.
 
-::: {#tab:examples}
 +--------------------------------------+
 | **Force-controlled Scheme**          |
 +--------------------------------------+
