### 2024-02-01 #5727 @kyleniemeyer
Hello @pprachas, I am doing some final checks before accepting.

In the paper, could you check/fix a few things?
- On page 3, in "215° and 180° archs", should that be "arcs" instead?
- I made a number of small edits in https://github.com/pprachas/fenics_arclength/pull/13, please review and mere.
