### 2024-02-01 #5727 @pprachas
Hi @kyleniemeyer,
1. Thanks for catching this! We meant "arches" -- we also fixed the table of examples to denote 215/180 arch instead of arc.
2. The branches are merged!

Thanks for all the suggestions and reformatting we really appreciate it! Please let us know if there are any more changes we need to do.
