### 2024-02-01 a5c39eb
Update paper.md - fixing table formatting.
